# public/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "🚗 Auto-École API - Authentification Authentik",
        "description": "API complète pour la gestion de l'authentification avec Authentik comme IAM central. Cette API permet l'inscription, la connexion, la gestion des utilisateurs et des tokens avec un système de contournement du Password Grant Flow.",
        "contact": {
            "name": "Support Auto-École API",
            "url": "https://auto-ecole-api.com",
            "email": "support@auto-ecole-api.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement local"
        },
        {
            "url": "https://9c8r7bbvybn.preview.infomaniak.website",
            "description": "Serveur de production Infomaniak"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement local"
        },
        {
            "url": "https://9c8r7bbvybn.preview.infomaniak.website",
            "description": "Serveur de production Infomaniak"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement local"
        },
        {
            "url": "https://9c8r7bbvybn.preview.infomaniak.website",
...
        },
        {
...
        },
...
    "paths": {
...
        "/api/documents": {
            "get": {
                "tags": [
                    "📄 Documents"
                ],
                "summary": "📋 Liste des documents",
                "operationId": "getDocuments",
                "parameters": [
                    {
                        "name": "dossier_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "valide",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "✅ Liste des documents"
                    }
                }
            },
            "post": {
                "tags": [
                    "📄 Documents"
                ],
                "summary": "➕ Créer un document",
                "operationId": "storeDocument",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "dossier_id",
                                    "type_document_id",
                                    "nom_fichier",
                                    "chemin_fichier",
                                    "type_mime",
                                    "taille_fichier"
                                ],
                                "properties": {
                                    "dossier_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "type_document_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "nom_fichier": {
                                        "type": "string",
                                        "example": "carte_identite.pdf"
                                    },
                                    "chemin_fichier": {
                                        "type": "string",
                                        "example": "/uploads/documents/carte_identite.pdf"
                                    },
                                    "type_mime": {
                                        "type": "string",
                                        "example": "application/pdf"
                                    },
                                    "taille_fichier": {
                                        "type": "integer",
                                        "example": 1024000
                                    },
                                    "valide": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "commentaires": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "✅ Document créé"
                    }
                }
            }
        },
...
        "/api/dossiers": {
            "get": {
                "tags": [
                    "📁 Dossiers"
                ],
                "summary": "📋 Liste des dossiers",
                "operationId": "getDossiers",
                "parameters": [
                    {
                        "name": "candidat_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "auto_ecole_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "en_attente",
                                "en_cours",
                                "valide",
                                "rejete"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "✅ Liste des dossiers"
                    }
                }
            },
            "post": {
                "tags": [
                    "📁 Dossiers"
                ],
                "summary": "➕ Créer un dossier",
                "operationId": "storeDossier",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "candidat_id",
                                    "auto_ecole_id",
                                    "formation_id",
                                    "date_creation"
                                ],
                                "properties": {
                                    "candidat_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "auto_ecole_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "formation_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "statut": {
                                        "type": "string",
                                        "enum": [
                                            "en_attente",
                                            "en_cours",
                                            "valide",
                                            "rejete"
                                        ]
                                    },
                                    "date_creation": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "commentaires": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "✅ Dossier créé"
                    }
                }
            }
        },
        "/api/dossiers/{id}": {
            "get": {
                "tags": [
                    "📁 Dossiers"
                ],
                "summary": "🔍 Détails d'un dossier",
                "operationId": "getDossier",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "✅ Détails du dossier"
                    }
                }
            },
            "put": {
                "tags": [
                    "📁 Dossiers"
                ],
                "summary": "✏️ Mettre à jour un dossier",
                "operationId": "updateDossier",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "statut": {
                                        "type": "string"
                                    },
                                    "commentaires": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "✅ Dossier mis à jour"
                    }
                }
            },
            "delete": {
                "tags": [
                    "📁 Dossiers"
                ],
                "summary": "🗑️ Supprimer un dossier",
                "operationId": "deleteDossier",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "✅ Dossier supprimé"
                    }
                }
            }
        },
        "/api/dossiers/{id}/upload-document": {
            "post": {
                "tags": [
                    "📁 Dossiers"
                ],
                "summary": "📤 Uploader un document",
                "description": "Upload un document et l'associe au dossier",
                "operationId": "uploadDocument",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "type_document_id",
                                    "fichier"
                                ],
                                "properties": {
                                    "type_document_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "fichier": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "commentaires": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "✅ Document uploadé avec succès"
                    },
                    "403": {
                        "description": "❌ Ce dossier ne vous appartient pas"
                    },
                    "404": {
                        "description": "❌ Dossier non trouvé"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Token d'accès JWT obtenu via l'authentification",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "bearer_token": {
                "type": "http",
                "description": "Token d'authentification JWT obtenu via /api/auth/login-direct",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}